Expose plural_login config to frontend
Also cleaned up a lot of unorganized gql code to match the patterns in the plural repo that have been working well.
